commit_211129

+++ src/controllers/videoController.js
@@ -1,5 +1,6 @@
 // import Video, { formatHashtags } from "../models/Video";
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async(req, res) => {        
         const videos = await Video.find({}).sort({ creatAt : `desc`});
@@ -8,37 +9,47 @@ export const home = async(req, res) => {
 }
 
 export const watch = async (req, res) => {
-    const id = req.params.id;
-    const video = await Video.findById(id);
+    const id = req.params.id;   
+    const video = await Video.findById(id).populate("owner");
+    console.log(video);
+    // const owner = await User.findById(video.owner);
     if(video){
         return res.render("video/watch", { pageTitle: video.title, video});
     } 
-    return res.render("404", { pageTitle: "Video Not Found.!"});    
+    return res.render("template/404", { pageTitle: "Video Not Found.!"});    
 };
 
 export const getEdit = async(req, res) => {
+    const { user: {_id}} = req.session
     const { id } = req.params;
     const video = await Video.findById(id);
     if (!video){
         return res.status(404).render("404", { pageTitle: "Video Not Found.!"});
     }
+    if (String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     return res.render("video/edit", { pageTitle:`Edit : ${video.title}`, video});     
 };
 
 export const postEdit = async (req, res) => {
+    const { user: {_id}} = req.session
     const { id } = req.params;   
     const { title, description, hashtags } = req.body;
     const video = await Video.findById(id);
     // const video = await Video.exists({ _id: id});
     if (!video){
-        return res.status(404).render("video/404", { pageTitle: "Video Not Found.!"});
-    } else {
+        return res.status(404).render("template/404", { pageTitle: "Video Not Found.!"});
+    } 
+    if (String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    } 
     await Video.findByIdAndUpdate(id,{
         title,
         description, 
         hashtags: Video.formatHashtags(hashtags),
         });
-    }       
+          
     // else{
     //     video.title = title;
     //     video.description = description;
@@ -53,6 +64,7 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async(req, res) => {
+    const {user: {_id}} = req.session;
     // const file = req.file;
     const { path } = req.file;
     const { title, description, hashtags } = req.body;
@@ -71,12 +83,16 @@ export const postUpload = async(req, res) => {
     */
    // Video db저장 create 방식 
    try {
-    await Video.create({
+    const newVideo = await Video.create({
         videoPath : path,
         title: title,
         description: description,
+        owner: _id,
         hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect(`/`);
    } catch(error) {
         return res.status(400).render("video/upload", {pageTitle: "UpLoad Video", errorMessage: error.message,});   
@@ -85,10 +101,18 @@ export const postUpload = async(req, res) => {
 };
 
 export const getDelete = async(req, res) => {
+    const { user: {_id}} = req.session
     const { id } = req.params;
+    const video = await Video.findById(id);
+    if (!video){
+        return res.status(404).render("template/404", { pageTitle: "Video Not Found.!"});
+    }
+    if (String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    } 
     await Video.findByIdAndDelete(id);
     return res.redirect(`/`);
-}
+};
 
 export const search = async(req, res) => {
     const { keyword } = req.query;
